refactor: organize topics messages under messages mod
Previously the Subscription message and subscription data was all in
the `topic_client` mod. To parallel the cache mod, we organize that
under `messages`.
