+++ src/topics/mod.rs
@@ -1,3 +1,5 @@
-mod topic_client;
+mod messages;
+pub use messages::subscription::{IntoTopicValue, Subscription, SubscriptionState};
 
+mod topic_client;
 pub use topic_client::TopicClient;
